move activity graph to new file
